bom and silkscreen updates
bom information added to all the parts and the silkscreen was all normalized to 40mils and placed so there is no more overlapping. a few designators were missing from the board so i brought them back. the only copper change was c1. the old footprint didn't match the part used on the pdf bom.  i modified it to match the designated part.